Develop

+++ src/main/java/com/example/carrotdiary/diary/controller/DiaryController.java
@@ -4,8 +4,6 @@ import com.example.carrotdiary.diary.dto.DiaryRequestDto;
 import com.example.carrotdiary.diary.service.DiaryService;
 import com.example.carrotdiary.global.common.Result;
 import com.example.carrotdiary.global.jwt.JwtUtils;
-import com.example.carrotdiary.image.entity.Image;
-import com.example.carrotdiary.image.service.ImageService;
 import jakarta.servlet.http.HttpServletRequest;
 import java.io.IOException;
 import java.util.List;
@@ -15,6 +13,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
+import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
@@ -26,28 +25,25 @@ import org.springframework.web.multipart.MultipartFile;
 public class DiaryController {
 
     private final DiaryService diaryService;
-    private final ImageService imageService;
     private final JwtUtils jwtUtils;
 
-    @PostMapping("/diary")
-    public ResponseEntity<Result> createDiary(HttpServletRequest req,
-                                         @ModelAttribute DiaryRequestDto.createDiaryDto diaryDto,
-                                         @RequestParam List<MultipartFile> images) throws IOException {
+    @PostMapping("/diary/{postDiaryId}")
+    public ResponseEntity<Long> createDiary(HttpServletRequest req, @PathVariable Long postDiaryId,
+                                            @ModelAttribute DiaryRequestDto.createDiaryDto diaryDto,
+                                            @RequestParam List<MultipartFile> images) throws IOException {
 
         String userEmail = jwtUtils.getUserEmail(req);
         if (userEmail == null) {
             throw new IllegalArgumentException("need login");
         }
 
-        List<Image> uploadImages = imageService.uploadImages(images);
+        Long diaryId = diaryService.createDiary(postDiaryId, diaryDto.getContent(), images);
 
-        Result diary = diaryService.createDiary(diaryDto.getPostDiaryId(), diaryDto.getContent(), uploadImages);
-
-        return ResponseEntity.status(HttpStatus.CREATED).body(diary);
+        return ResponseEntity.status(HttpStatus.CREATED).body(diaryId);
 
     }
 
-    @GetMapping("/diary/{postDiaryId}")
+    @GetMapping("/diary/all/{postDiaryId}")
     public ResponseEntity<Result> getAllDiary(HttpServletRequest req, @PathVariable("postDiaryId") Long postDiaryId) {
 
         String userEmail = jwtUtils.getUserEmail(req);
@@ -61,7 +57,7 @@ public class DiaryController {
 
     }
 
-    @GetMapping("/diary/{diaryId}")
+    @GetMapping("/diary/detail/{diaryId}")
     public ResponseEntity<Result> getDiary(HttpServletRequest req, @PathVariable("diaryId") Long diaryId) {
 
         String userEmail = jwtUtils.getUserEmail(req);
@@ -74,27 +70,26 @@ public class DiaryController {
         return ResponseEntity.ok(diary);
     }
 
-    @GetMapping("diary/{postId}")
-    public ResponseEntity<Result> getPreviewDiary(HttpServletRequest req, @PathVariable("postId") Long postId) {
-
+    @PatchMapping("/diary/{diaryId}")
+    public ResponseEntity<Result> updateDiary(HttpServletRequest req, @PathVariable("diaryId") Long diaryId,
+                                              @ModelAttribute DiaryRequestDto.updateDiaryDto updateDiaryDto,
+                                              @RequestParam(required = false) List<MultipartFile> images) throws IOException {
         String userEmail = jwtUtils.getUserEmail(req);
         if (userEmail == null) {
             throw new IllegalArgumentException("need login");
         }
 
-        Result previewDiary = diaryService.getPreviewDiary(postId);
+        Result result = diaryService.updateDiary(diaryId, updateDiaryDto, images);
 
-        return ResponseEntity.ok(previewDiary);
+        return ResponseEntity.ok(result);
     }
-
-//    @PatchMapping("/diary/{diaryId}")
-//    public ResponseEntity<?> updateDiary(@PathVariable("diaryId") Long diaryId, DiaryRequestDto) {
-//
-//    } 이미지 수정 부분 생각
 
     @DeleteMapping("/diary/{diaryId}")
     public ResponseEntity<String> deleteDiary(HttpServletRequest req, @PathVariable("diaryId") Long diaryId) {
         String userEmail = jwtUtils.getUserEmail(req);
+        if (userEmail == null) {
+            throw new IllegalArgumentException("need login");
+        }
 
         diaryService.deleteDiary(diaryId);
